Move podNIC code to podnic.go
The current go file podinterface.go contains Phase1/Phase2 code and the
libvirt spec domain generator, this change move the Phase1/Phase2
(podNIC) to podnic.go
